requirements.txt
mcp
fastmcp
python-dotenv
markitdown[all]
